refactor(settings): move from route to dialog (#235)
