+++ src/lib/components/settings/Category.svelte
