+++ src/lib/components/settings/SidebarActions.svelte
