+++ src/lib/components/settings/categories/appearance.ts
