+++ src/lib/components/settings/categories/highlights.ts
