+++ src/lib/components/settings/custom/ClearCache.svelte
