+++ src/lib/components/settings/custom/highlights/Color.svelte
