+++ src/lib/components/settings/custom/highlights/Edit.svelte
