+++ src/lib/components/settings/custom/highlights/Keyword.svelte
