+++ src/lib/components/settings/custom/highlights/Style.svelte
