--- src/routes/settings/+page.svelte
@@ -1,98 +1,0 @@
-<script lang="ts">
-	import { Tabs } from "bits-ui";
-	import { beforeNavigate, goto } from "$app/navigation";
-	import { app } from "$lib/app.svelte";
-	import Button from "$lib/components/ui/Button.svelte";
-	import Separator from "$lib/components/ui/Separator.svelte";
-	import { log } from "$lib/log";
-	import { settings } from "$lib/settings";
-	import { logOut } from "$lib/twitch/auth";
-	import SignOut from "~icons/ph/sign-out";
-	import X from "~icons/ph/x";
-	import Category from "./Category.svelte";
-	import SidebarActions from "./SidebarActions.svelte";
-	import type { SettingsCategory } from "./types";
-
-	const imports = import.meta.glob<SettingsCategory>(
-		["./categories/*.ts", "!./categories/util.ts"],
-		{ eager: true, import: "default" },
-	);
-
-	const categories = Object.values(imports).toSorted((a, b) => a.order - b.order);
-
-	beforeNavigate(async () => {
-		await settings.saveNow();
-		log.info("Settings saved");
-	});
-</script>
-
-<Tabs.Root
-	id="settings-tabs"
-	class="relative flex grow overflow-hidden"
-	orientation="vertical"
-	value={categories[0].label}
->
-	<div
-		class="h-full w-44 shrink-0 p-2 transition-[width] duration-300 ease-out-quint md:w-64 lg:w-72"
-	>
-		<Tabs.List class="space-y-0.5">
-			{#each categories as category (category.label)}
-				<Tabs.Trigger value={category.label}>
-					{#snippet child({ props })}
-						<Button
-							class="text-muted-foreground data-[state=active]:bg-muted data-[state=active]:text-foreground"
-							variant="ghost"
-							{...props}
-						>
-							<category.icon />
-							{category.label}
-						</Button>
-					{/snippet}
-				</Tabs.Trigger>
-			{/each}
-		</Tabs.List>
-
-		<Separator />
-
-		<div class="space-y-1">
-			<SidebarActions />
-		</div>
-
-		<Separator />
-
-		<Button class="text-muted-foreground" variant="ghost" onclick={logOut}>
-			<SignOut />
-			<span class="text-sm">Log out</span>
-		</Button>
-	</div>
-
-	<div class="relative grow overflow-y-auto border-l bg-accent/15 p-4 pb-8">
-		<Button
-			class="absolute top-2 right-2"
-			size="icon"
-			variant="ghost"
-			onclick={() => (app.history.canGoBack ? app.history.back() : goto("/"))}
-			aria-label="Close settings"
-		>
-			<X />
-		</Button>
-
-		{#each categories as category (category.label)}
-			<Tabs.Content class="mx-auto max-w-xl" value={category.label}>
-				<Category {category} />
-			</Tabs.Content>
-		{/each}
-	</div>
-</Tabs.Root>
-
-<style>
-	:global(#settings-tabs [data-slot="separator"]) {
-		margin: calc(var(--spacing) * 2) 0;
-	}
-
-	div:first-child :global(button) {
-		width: 100%;
-		display: flex;
-		justify-content: flex-start;
-	}
-</style>
